Various frontend cleanups (#141)
* Split util.ts across util/urls.ts + util/format.ts

* Drop redundant classes in app.css
