+++ client/src/app.css
@@ -3,46 +3,3 @@
 mark {
   background-color: var(--background-contrast-info);
 }
-
-/* icons straight from http://remixicon.com/ (not included in the DSFR) */
-/* The following are downloaded from http://remixicon.com/, and modified to
- * remove their width and height attributes so they're customizable through css */
-.fr-fi-bank-line {
-  background-color: var(--text-action-high-blue-france);
-  mask: url(../icons/bank-line.svg);
-}
-
-.fr-fi-map-2-line {
-  background-color: var(--text-action-high-blue-france);
-  mask: url(../icons/map-2-line.svg);
-}
-
-.fr-fi-calendar-check-line {
-  background-color: var(--text-action-high-blue-france);
-  mask: url(../icons/calendar-check-line.svg);
-}
-
-.fr-fi-cake-2-line {
-  background-color: var(--text-action-high-blue-france);
-  mask: url(../icons/cake-2-line.svg);
-}
-
-/* aside on the "fiche de données" page */
-.aside-entry {
-  align-items: center;
-  display: flex;
-  gap: 10px;
-  margin-bottom: 2em;
-}
-
-.aside-entry [class*="fr-fi"],
-.aside-entry [class*="fr-icon"] {
-  color: var(--text-action-high-blue-france);
-  display: inline-block;
-  height: 32px;
-  width: 32px;
-}
-
-.aside-entry p {
-  margin-bottom: 0;
-}
