+++ client/src/lib/util/format.ts
@@ -1,0 +1,10 @@
+export const pluralize = (
+  amount: number,
+  singleText: string,
+  pluralText: string
+) => {
+  if (amount === 1) {
+    return singleText;
+  }
+  return pluralText;
+};
