+++ client/src/lib/util/urls.ts
@@ -1,14 +1,3 @@
-export const pluralize = (
-  amount: number,
-  singleText: string,
-  pluralText: string
-) => {
-  if (amount === 1) {
-    return singleText;
-  }
-  return pluralText;
-};
-
 /**
  * Create an URL query string, including a leading "?" if needed, dropping any null or undefined values.
  */
